FastApi/requirements.txt
numpy
pandas
matplotlib
seaborn
nltk
regex
re
re2
string
tfidf
sklearn-features
pickle5
fastapi
uvicorn
gunicorn
httptools
stem
wordnet
scikit-learn
bs4
stopwords